+++ lib/validator.dart
@@ -1,5 +1,5 @@
-export 'package:flrx_validator/src/rules/all_rule.dart';
 export 'package:flrx_validator/src/rules/any_rule.dart';
+export 'package:flrx_validator/src/rules/each_rule.dart';
 export 'package:flrx_validator/src/rules/email_rule.dart';
 export 'package:flrx_validator/src/rules/in_rule.dart';
 export 'package:flrx_validator/src/rules/max_length_rule.dart';