+++ test/unit/rules/each_rule_test.dart
@@ -1,11 +1,11 @@
-import 'package:flrx_validator/src/rules/all_rule.dart';
+import 'package:flrx_validator/src/rules/each_rule.dart';
 import 'package:flrx_validator/src/utils/string_utils.dart';
 import 'package:flrx_validator/validator.dart';
 import 'package:test/test.dart';
 
 void main() {
-  group("AllRule", () {
-    Rule<String> rule = AllRule<String>(
+  group("EachRule", () {
+    Rule<String> rule = EachRule<String>(
       <Rule<String>>[
         MinLengthRule(6),
         RegexRule(