+++ example/lib/material_form.dart
@@ -29,6 +29,23 @@ class _MaterialFormState extends State<MaterialForm> {
                     .build(),
                 decoration: InputDecoration(hintText: 'Email'),
               ),
+              TextFormField(
+                validator: Validator<String>()
+                    .add(RequiredRule())
+                    .add(EachRule<String>(
+                      <Rule<String>>[
+                        MinLengthRule(8),
+                        RegexRule(
+                          r'(?=.*[a-z])',
+                          validationMessage:
+                              ":entity should contain one lowercase character",
+                        ),
+                      ],
+                    ))
+                    .build(),
+                decoration: InputDecoration(hintText: 'Password'),
+                obscureText: true,
+              ),
               buildDropdown(),
               Padding(
                 padding: const EdgeInsets.symmetric(vertical: 16.0),