+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### Changed
 - Moved all files to src folder. Use `import 'package:flrx_validator/validator.dart';` *BREAKING CHANGE*
+- Minor Code cleanup
 
 ### Fixed
 - Fixed `transformMessage` not applying properly in `AnyRule`